[SPARK-51921][SS][PYTHON][4.0] Use long type for TTL duration in millisecond in transformWithState
### What changes were proposed in this pull request?

- Use long type (int64 for protobuf) for TTL duration in millisecond
- Fixed `test_transform_with_map_state_metadata` test case.

### Why are the changes needed?

Allow users to set longer period of time for TTL duration.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Updated test case

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50745 from bogao007/ttl-4.0.

Authored-by: bogao007 <bo.gao@databricks.com>

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/python/streaming/TransformWithStateInPandasStateServer.scala
@@ -636,7 +636,7 @@ class TransformWithStateInPandasStateServer(
       stateName: String,
       schemaString: String,
       stateType: StateVariableType.StateVariableType,
-      ttlDurationMs: Option[Int],
+      ttlDurationMs: Option[Long],
       mapStateValueSchemaString: String = null): Unit = {
     val schema = StructType.fromString(schemaString)
     val expressionEncoder = ExpressionEncoder(schema).resolveAndBind()
